pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "local-token-expansion"
version = "0.1.0"
description = "Local Token Expansion (LTE) benchmark runner for MLX models"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Local Token Expansion contributors" }]
dependencies = [
  "PyYAML>=6.0",
  # Only installed on Apple Silicon macOS; CI remains MLX-free.
  "mlx-lm>=0.20.0; platform_system == 'Darwin' and platform_machine == 'arm64'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.4.8",
]

[project.scripts]
lte = "lte.cli:main"

[tool.setuptools]
packages = ["lte", "lte.backends"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]
